+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Current (in progress)
+
+- Nothing yet
+
 ## 10.5.0 (2025-07-02)
 
 - Allow temporal intervals with no start date [#3353](https://github.com/opendatateam/udata/pull/3353)